SS Emotions EndPoints

+++ services/emotions.js
@@ -1,8 +1,9 @@
+const emotionTypeModel = require('../database/emotionTypesModel')
 const emotionsModel = require('../database/emotionsModel')
-const emotionsEntryModel = require('../database/emotionsEntryModel')
+const stringUtils = require('../utils/stringUtils')
 const { Op } = require('sequelize')
 
-async function create(description, intensity_weight, type_id, user_id){
+async function create(user_id, description, intensity_weight, type_id) {
     emotionsModel.create({
         description: description,
         intensity_weight: intensity_weight,
@@ -11,47 +12,75 @@ async function create(description, intensity_weight, type_id, user_id){
     })
 }
 
-async function remove(emotion_id){
-    emotionsModel.destroy({
-        where:{
-            id: emotion_id
-        }
-    })
+async function edit(user_id, emotion_id, description, intensity_weight, type_id) {
+    if (!isNaN(emotion_id)) {
+        return await emotionsModel.findOne({
+            where: {
+                id: emotion_id,
+                user_id: {
+                    [Op.eq]: user_id,
+                    [Op.and]: { [Op.ne]: undefined }
+                }
+            }
+        }).then(emotion => {
+            if(!stringUtils.isNullOrEmpty(description)){
+                emotion.description = description
+            }
+
+            if(intensity_weight != undefined && !isNaN(intensity_weight)){
+                emotion.intensity_weight = intensity_weight
+            }
+
+            if(type_id != undefined && !isNaN(type_id)){
+                emotion.type_id = type_id
+            }
+        })
+    }
 }
 
-async function getAll(user_id){
-    emotionsModel.findAll({
+async function remove(user_id, emotion_id) {
+    if (!isNaN(emotion_id)) {
+        return await emotionsModel.destroy({
+            where: {
+                id: emotion_id,
+                user_id: user_id,
+            }
+        }).then(delete_count => {
+            return delete_count > 0
+        })
+    }
+    return false
+}
+
+async function getAll(user_id) {
+    return await emotionsModel.findAll({
         raw: true,
         where: {
             user_id: user_id
         }
-    }).then(emotions => {
-        return emotions
     })
 }
 
-async function get(emotion_id){
-    emotionsModel.findOne({
+async function get(user_id, emotion_id) {
+    return await emotionsModel.findOne({
         raw: true,
         where: {
-            emotion_id: emotion_id
+            id: emotion_id,
+            user_id: {
+                [Op.or]:
+                    [
+                        { user_id: undefined },
+                        { user_id: user_id },
+                    ]
+            }
         }
-    }).then(emotion => {
-        return emotion
-    })
-}
-
-async function addEmotionToEntry(emotion_id, entry_id) {
-    emotionsEntryModel.create({
-        emotion_id: emotion_id,
-        entry_id: entry_id
     })
 }
 
 module.exports = {
     create,
+    edit,
     remove,
-    getAll,
     get,
-    addEmotionToEntry
+    getAll,
 }
